+++ go.mod
@@ -1,0 +1,5 @@
+module nba-result
+
+go 1.14
+
+require github.com/gizak/termui/v3 v3.1.0
